requirements.txt
# RAG-Benchmark Hauptprojekt Dependencies
#
# Author: Lukas Schaumlöffel
# Master Informatik (HAW Hamburg)

# Core Data Science Stack
numpy>=2.3.2,<3.0.0
pandas>=2.3.1,<3.0.0
scipy>=1.14.1,<2.0   ; python_version >= "3.13"

# Machine Learning & NLP
scikit-learn>=1.6.1,<2.0 ; python_version >= "3.13"
scikit-learn>=1.3,<2.0   ; python_version < "3.13"
sentence-transformers>=2.2.0,<3.0.0
spacy>=3.7.0,<4.0.0

# Vector Database & Search
faiss-cpu>=1.7.0,<2.0.0

# Graph Database
neo4j>=5.0.0,<6.0.0

# LLM Integration (LangChain Architecture)
langchain>=0.3.27,<0.4.0
langchain-openai>=0.1.0,<1.0.0
langchain-core>=0.2.0,<1.0.0
openai>=1.99.1,<2.0.0

# Evaluation Metrics
nltk>=3.8.0,<4.0.0
rouge-score>=0.1.2,<1.0.0

# Configuration & Validation
pyyaml>=6.0.0,<7.0.0
pydantic>=2.0.0,<3.0.0

# Visualization & Reporting
matplotlib>=3.7.0,<4.0.0
seaborn>=0.12.0,<1.0.0

# System Utilities
psutil>=5.9.0,<6.0.0
tqdm>=4.65.0,<5.0.0

# Development & Jupyter (optional but recommended)
jupyter>=1.0.0,<2.0.0
ipykernel>=6.30.0,<7.0.0

# Text Processing Utilities
unidecode>=1.3.0,<2.0.0
python-dotenv>=1.0.0,<2.0.0

# Performance & Caching
joblib>=1.3.0,<2.0.0
diskcache>=5.6.0,<6.0.0

# Additional NLP Models (to be downloaded separately)
# python -m spacy download de_core_news_sm
# python -m spacy download en_core_web_sm
# python -c "import nltk; nltk.download('punkt'); nltk.download('stopwords')"